Changes due to the review
