+++ .vscode/settings.json
@@ -148,6 +148,7 @@
         "ICONEU",
         "idxmax",
         "IGRA",
+        "Ilya",
         "imageio",
         "imread",
         "imshow",
@@ -156,15 +157,18 @@
         "Interquartile",
         "intp",
         "ipynb",
+        "ipython",
         "ipywidgets",
         "isbijective",
         "isin",
+        "isort",
         "ivar",
         "jsonpickle",
         "Junqueira",
         "jupyter",
         "Kaleb",
         "Karman",
+        "labelrotation",
         "linalg",
         "linestyle",
         "linewidth",
@@ -212,6 +216,7 @@
         "noaaruc",
         "noaarucsounding",
         "num2pydate",
+        "numericalunits",
         "numfig",
         "numpy",
         "numref",
@@ -221,6 +226,7 @@
         "polystyle",
         "powerseries",
         "Prandtl",
+        "prettytable",
         "Projeto",
         "prometheus",
         "pydata",
@@ -228,11 +234,14 @@
         "PYPI",
         "pyplot",
         "pyproject",
+        "pytest",
         "pytz",
+        "quantile",
         "Rdot",
         "referece",
         "relativetoground",
         "repr",
+        "reversesort",
         "reynolds",
         "rightarrow",
         "ROABs",
@@ -252,6 +261,7 @@
         "setrail",
         "simplekml",
         "SIRGAS",
+        "Sobol",
         "solidmotor",
         "somgl",
         "Somigliana",
@@ -260,9 +270,12 @@
         "SRTM",
         "SRTMGL",
         "Stano",
+        "statsmodels",
         "STFT",
         "subintervals",
         "suptitle",
+        "supxlabel",
+        "supylabel",
         "ticklabel",
         "timedelta",
         "timezonefinder",
