+++ rocketpy/sensitivity/sensivity_model.py
@@ -1,8 +1,8 @@
-import matplotlib.pyplot as plt
 import numpy as np
-from scipy.stats import norm
 
-from ..tools import check_requirement_version, import_optional_dependency
+from rocketpy.plots.sensitivity_plots import _SensitivityModelPlots
+from rocketpy.prints.sensitivity_prints import _SensitivityModelPrints
+from rocketpy.tools import check_requirement_version, import_optional_dependency
 
 
 class SensitivityModel:
@@ -80,13 +80,10 @@ class SensitivityModel:
 
         self._fitted = False
 
-        return
+        self.prints = _SensitivityModelPrints(self)
+        self.plots = _SensitivityModelPlots(self)
 
-    def set_parameters_nominal(
-        self,
-        parameters_nominal_mean,
-        parameters_nominal_sd,
-    ):
+    def set_parameters_nominal(self, parameters_nominal_mean, parameters_nominal_sd):
         """Set parameters nominal mean and standard deviation
 
         Parameters
@@ -105,13 +102,13 @@ class SensitivityModel:
         """
         if len(parameters_nominal_mean) != self.n_parameters:
             raise ValueError(
-                "Nominal mean array length does not match number of \
-                 parameters passed at initialization."
+                "Nominal mean array length does not match number of "
+                "parameters passed at initialization."
             )
         if len(parameters_nominal_sd) != self.n_parameters:
             raise ValueError(
-                "Nominal sd array length does not match number of parameters \
-                 passed at initialization."
+                "Nominal sd array length does not match number of parameters "
+                "passed at initialization."
             )
 
         for i, parameter in enumerate(self.parameters_names):
@@ -120,12 +117,7 @@ class SensitivityModel:
 
         self._nominal_parameters_passed = True
 
-        return
-
-    def set_target_variables_nominal(
-        self,
-        target_variables_nominal_value,
-    ):
+    def set_target_variables_nominal(self, target_variables_nominal_value):
         """Set target variables nominal value (mean)
 
         Parameters
@@ -155,12 +147,7 @@ class SensitivityModel:
 
         self._nominal_target_passed = True
 
-        return
-
-    def _estimate_parameter_nominal(
-        self,
-        parameters_matrix,
-    ):
+    def _estimate_parameter_nominal(self, parameters_matrix):
         """Estimates parameters nominal values
 
         Parameters
@@ -182,12 +169,7 @@ class SensitivityModel:
                 parameters_matrix[:, i]
             )
 
-        return
-
-    def _estimate_target_nominal(
-        self,
-        target_data,
-    ):
+    def _estimate_target_nominal(self, target_data):
         """Estimates target variables nominal values
 
         Parameters
@@ -213,13 +195,7 @@ class SensitivityModel:
                     target_data[:, i]
                 )
 
-        return
-
-    def fit(
-        self,
-        parameters_matrix,
-        target_data,
-    ):
+    def fit(self, parameters_matrix, target_data):
         """Fits sensitivity model
 
         Parameters
@@ -236,14 +212,11 @@ class SensitivityModel:
         -------
         None
         """
-        # imports statsmodels for OLS method
         sm = import_optional_dependency("statsmodels.api")
 
-        # Checks if data is in conformity with initialization info
         self.__check_conformity(parameters_matrix, target_data)
 
-        # If nominal parameters are not set previous to fit, then we
-        # must estimate them
+        # If nominal parameters are not set previous to fit, we must estimate them
         if not self._nominal_parameters_passed:
             self._estimate_parameter_nominal(parameters_matrix)
         if not self._nominal_target_passed:
@@ -268,7 +241,11 @@ class SensitivityModel:
         if target_data.ndim == 1:
             target_data = target_data.reshape(self.number_of_samples, 1)
 
-        # Estimation
+        self.__estimation_loop(target_data, sm, parameters_sd, X)
+
+        self._fitted = True
+
+    def __estimation_loop(self, target_data, sm, parameters_sd, X):
         for i, target_variable in enumerate(self.target_variables_names):
             nominal_value = self.target_variables_info[target_variable]["nominal_value"]
             Y = np.array(target_data[:, i] - nominal_value)
@@ -279,206 +256,30 @@ class SensitivityModel:
             # Compute sensitivity
             beta = fitted_model.params[1:]  # skipping the intercept
             var_eps = fitted_model.scale
-            var_Y = var_eps
+            var_y = var_eps
             for k, parameter in enumerate(self.parameters_names):
                 sensitivity = np.power(beta[k], 2) * np.power(parameters_sd[k], 2)
                 self.target_variables_info[target_variable]["sensitivity"][
                     parameter
                 ] = sensitivity
-                var_Y += sensitivity
+                var_y += sensitivity
 
-            self.target_variables_info[target_variable]["var"] = var_Y
-            self.target_variables_info[target_variable]["sd"] = np.sqrt(var_Y)
+            self.target_variables_info[target_variable]["var"] = var_y
+            self.target_variables_info[target_variable]["sd"] = np.sqrt(var_y)
 
             for k, parameter in enumerate(self.parameters_names):
                 self.target_variables_info[target_variable]["sensitivity"][
                     parameter
-                ] /= var_Y
+                ] /= var_y
             self.target_variables_info[target_variable]["LAE"] = var_eps
-            self.target_variables_info[target_variable]["LAE"] /= var_Y
+            self.target_variables_info[target_variable]["LAE"] /= var_y
 
-        self._fitted = True
-        return
+    def info(self):
+        self.prints.all()
 
-    def plot(self, target_variable="all"):
-        """Creates barplot showing the sensitivity of the target_variable due
-        to parameters
-
-        Parameters
-        ----------
-        target_variable : str, optional
-            Name of the target variable used to show sensitivity. It can also
-            be "all", in which case a plot is created for each target variable
-            in which the model was fitted. The default is "all".
-
-        Returns
-        -------
-        None
-        """
-        self.__check_if_fitted()
-
-        if (target_variable not in self.target_variables_names) and (
-            target_variable != "all"
-        ):
-            raise ValueError(
-                f"Target variable {target_variable} was not listed in \
-                  initialization!"
-            )
-
-        # Parameters bars are blue colored
-        # LAE bar is red colored
-        bar_colors = self.n_parameters * ["blue"]
-        bar_colors.append("red")
-
-        if target_variable == "all":
-            for i in range(self.n_target_variables):
-                fig, axs = plt.subplots()
-                fig.supxlabel("")
-                fig.supylabel("Sensitivity (%)")
-                x = [parameter for parameter in self.parameters_names]
-                x.append("LAE")
-                current_target_variable = self.target_variables_names[i]
-                y = np.empty(self.n_parameters + 1)
-                for j in range(self.n_parameters):
-                    parameter = x[j]
-                    y[j] = (
-                        100
-                        * self.target_variables_info[current_target_variable][
-                            "sensitivity"
-                        ][parameter]
-                    )
-                y[self.n_parameters] = (
-                    100 * self.target_variables_info[current_target_variable]["LAE"]
-                )
-                axs.bar(x, y, color=bar_colors)
-                axs.set_title(current_target_variable)
-                axs.tick_params(labelrotation=90)
-            plt.show()
-
-            return
-
-        fig, axs = plt.subplots()
-        fig.supxlabel("")
-        fig.supylabel("Sensitivity (%)")
-        x = [parameter for parameter in self.parameters_names]
-        x.append("LAE")
-        y = np.empty(self.n_parameters + 1)
-        for j in range(self.n_parameters):
-            parameter = x[j]
-            y[j] = (
-                100
-                * self.target_variables_info[target_variable]["sensitivity"][parameter]
-            )
-        y[self.n_parameters] = 100 * self.target_variables_info[target_variable]["LAE"]
-        axs.bar(x, y, color=bar_colors)
-        axs.set_title(target_variable)
-        axs.tick_params(labelrotation=90)
-        plt.show()
-
-        return
-
-    def summary(self, digits=4, alpha=0.95):
-        """Formats parameter sensitivity information in a prettytable
-        and prints it
-
-        Parameters
-        ----------
-        digits : int, optional
-            Number of decimal digits printed on tables, by default 4
-        alpha: float, optional
-            Significance level used for prediction intervals, by default 0.95
-
-        Returns
-        -------
-        None
-        """
-
-        self.__check_if_fitted()
-
-        pt = import_optional_dependency("prettytable")
-
-        if self._nominal_parameters_passed:
-            nominal_mean_text = "Nominal mean"
-            nominal_sd_text = "Nominal sd"
-        else:
-            nominal_mean_text = "Estimated mean"
-            nominal_sd_text = "Estimated sd"
-        for target_variable in self.target_variables_names:
-            model = self.target_variables_info[target_variable]["model"]
-            coef = model.params[1:]  # skipping intercept
-            pvalues = model.pvalues
-
-            sensitivity_table = pt.PrettyTable()
-            sensitivity_table.title = f"Summary {target_variable}"
-
-            sensitivity_table.field_names = [
-                "Parameter",
-                "Sensitivity (%)",
-                nominal_mean_text,
-                nominal_sd_text,
-                "Regression Coefficient",
-                "p-value",
-            ]
-
-            for i in range(self.n_parameters):
-                parameter = self.parameters_names[i]
-                beta = coef[i]
-                pval = pvalues[i]
-                sensitivity = self.target_variables_info[target_variable][
-                    "sensitivity"
-                ][parameter]
-                sensitivity_table.add_row(
-                    [
-                        parameter,
-                        round(100 * sensitivity, digits),
-                        round(self.parameters_info[parameter]["nominal_mean"], digits),
-                        round(self.parameters_info[parameter]["nominal_sd"], digits),
-                        round(beta, digits),
-                        round(pval, digits),
-                    ]
-                )
-            sensitivity_table.add_row(
-                [
-                    "Linear Approx. Error (LAE)",
-                    round(
-                        100 * self.target_variables_info[target_variable]["LAE"],
-                        digits,
-                    ),
-                    "",
-                    "",
-                    "",
-                    "",
-                ]
-            )
-            sensitivity_table.sortby = "Sensitivity (%)"
-            sensitivity_table.reversesort = True
-
-            print(sensitivity_table)
-
-            table = pt.PrettyTable()
-            nominal_value = round(
-                self.target_variables_info[target_variable]["nominal_value"], digits
-            )
-            norm_quantile = norm.ppf((1 + alpha) / 2)
-            if self._nominal_target_passed:
-                table.add_row([f"Nominal value: {nominal_value}"])
-            else:
-                table.add_row([f"Estimated value: {nominal_value}"])
-            target_sd = self.target_variables_info[target_variable]["sd"]
-            table.add_row([f"Std: {round(target_sd, digits)}"])
-            ci_lower = round(nominal_value - norm_quantile * target_sd, digits)
-            ci_upper = round(nominal_value + norm_quantile * target_sd, digits)
-            table.add_row(
-                [
-                    f"{round(100 * alpha, 0)}% Prediction Interval: [{ci_lower}, {ci_upper}]"
-                ]
-            )
-            column_width = len(sensitivity_table._hrule)
-            # Make tables borders match
-            table.field_names = [(column_width - 4) * " "]
-            print(table)
-
-        return
+    def all_info(self):
+        self.prints.all()
+        self.plots.all()
 
     def __check_conformity(
         self,
@@ -527,18 +328,15 @@ class SensitivityModel:
                  and target data."
             )
 
-        return
-
-    def __check_if_fitted(self):
-        """Checks if model is fitted
+    def _raise_error_if_not_fitted(self):
+        """Checks if model is fitted or raise an error.
 
         Returns
         -------
         None
         """
         if not self._fitted:
-            raise Exception("SensitivityModel must be fitted!")
-        return
+            raise ValueError("SensitivityModel must be fitted!")
 
     def __check_requirements(self):
         """Check if extra requirements are installed. If not, print a message
@@ -568,4 +366,3 @@ class SensitivityModel:
                 "Given the above errors, some methods may not work. Please run "
                 + "'pip install rocketpy[sensitivity]' to install extra requirements."
             )
-        return None
